feat : byte-string 변환부 util로 빼주었다.
