+++ src/main/java/utils/ByteStringUtil.java
@@ -1,0 +1,20 @@
+package utils;
+
+public class ByteStringUtil {
+    public static String byteArrayToHex(byte[] bytes) {
+        StringBuilder sb = new StringBuilder();
+        for (final byte b : bytes) {
+            sb.append(String.format("%02X", b & 0xff));
+        }
+        return sb.toString();
+    }
+    public static byte[] decodeWithHex(String hexStr) {
+        int len = hexStr.length();
+        byte[] data = new byte[len / 2];
+        for (int i = 0; i < len; i += 2) {
+            data[i / 2] = (byte) ((Character.digit(hexStr.charAt(i), 16) << 4)
+                    + Character.digit(hexStr.charAt(i + 1), 16));
+        }
+        return data;
+    }
+}
